% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_text.R
\name{byte_text_encoder}
\alias{byte_text_encoder}
\title{Byte Text Encoder}
\usage{
byte_text_encoder(additional_tokens = NULL)
}
\arguments{
\item{additional_tokens}{list of additional tokens. These will be assigned
vocab ids \link{1, 1+len(additional_tokens)}. Useful for things like "end-of-string"
tokens (e.g. "").}
}
\description{
Byte encodes text
}
\seealso{
\code{\link[=save_byte_text_encoder]{save_byte_text_encoder()}}, \code{\link[=load_byte_text_encoder]{load_byte_text_encoder()}}, \code{\link[=encode]{encode()}}
and \code{\link[=decode]{decode()}}
}
