% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_text.R
\name{decode}
\alias{decode}
\title{Decode text}
\usage{
decode(token_text_encoder, ids)
}
\arguments{
\item{token_text_encoder}{a Token Text Encoder created with \code{\link[=token_text_encoder]{token_text_encoder()}}}

\item{ids}{ids to decode using the Token Text encoder.}
}
\description{
Decode text
}
\seealso{
\code{\link[=token_text_encoder]{token_text_encoder()}}
}
