% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_text.R
\name{decode_text}
\alias{decode_text}
\title{Decode text}
\usage{
decode_text(token_text_encoder, ids)
}
\arguments{
\item{token_text_encoder}{a Token Text Encoder created with `token_text_encoder()`}

\item{ids}{ids to decode using the Token Text encoder.}
}
\description{
Decode text
}
