% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_text.R
\name{encode}
\alias{encode}
\title{Encode text}
\usage{
encode(token_text_encoder, text)
}
\arguments{
\item{token_text_encoder}{a Token Text Encoder created with \code{\link[=token_text_encoder]{token_text_encoder()}}}

\item{text}{string to encode using the Token Text encoder}
}
\description{
Encode text
}
\seealso{
\code{\link[=token_text_encoder]{token_text_encoder()}}
}
