% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_text.R
\name{encode_text}
\alias{encode_text}
\title{Encode text}
\usage{
encode_text(token_text_encoder, text)
}
\arguments{
\item{token_text_encoder}{a Token Text Encoder created with `token_text_encoder()`}

\item{text}{string to encode using the Token Text encoder}
}
\description{
Encode text
}
