% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_tfds}
\alias{install_tfds}
\title{Installs TensorFlow Datasets}
\usage{
install_tfds(method = c("auto", "virtualenv", "conda"), conda = "auto",
  version = "default", envname = NULL, restart_session = TRUE,
  conda_python_version = "3.6", ...)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows (as this isn't supported by TensorFlow). Note also
that since this command runs without privilege the "system" method is
available only on Windows.}

\item{conda}{Path to conda executable (or "auto" to find conda using the
PATH and other conventional install locations).}

\item{version}{By default the latest PyPi version will be installed.

Alternatively, you can provide the full URL to an installer binary (e.g.
for a nightly binary).}

\item{envname}{Name of Python environment to install within}

\item{restart_session}{Restart R session after installing (note this will
only occur within RStudio).}

\item{conda_python_version}{the python version installed in the created conda
environment. Python 3.6 is installed by default.}

\item{...}{other arguments passed to \code{\link[reticulate:conda_install]{reticulate::conda_install()}} or
\code{\link[reticulate:virtualenv_install]{reticulate::virtualenv_install()}}.}
}
\description{
Installs TensorFlow Datasets
}
