% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_text.R
\name{load_token_text_encoder}
\alias{load_token_text_encoder}
\title{Loads a Token Text Encoder}
\usage{
load_token_text_encoder(path)
}
\arguments{
\item{path}{path of a saved token text encoder.}
}
\description{
Loads a Token Text Encoder
}
