% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_text.R
\name{save_byte_text_encoder}
\alias{save_byte_text_encoder}
\title{Saves a byte text encoder to a file}
\usage{
save_byte_text_encoder(byte_text_encoder, path)
}
\arguments{
\item{byte_text_encoder}{a Byte Text Encoder created with \code{\link[=byte_text_encoder]{byte_text_encoder()}}}

\item{path}{path to save the text encoder.}
}
\description{
Saves a byte text encoder to a file
}
\seealso{
\code{\link[=byte_text_encoder]{byte_text_encoder()}}
}
