% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_text.R
\name{save_subword_text_encoder}
\alias{save_subword_text_encoder}
\title{Saves a subword text encoder to a file}
\usage{
save_subword_text_encoder(subword_text_encoder, path)
}
\arguments{
\item{subword_text_encoder}{a Subword Text Encoder created with \code{\link[=subword_text_encoder]{subword_text_encoder()}}}

\item{path}{path to save the text encoder.}
}
\description{
Saves a subword text encoder to a file
}
\seealso{
\code{\link[=byte_text_encoder]{byte_text_encoder()}}
}
