% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_text.R
\name{save_token_text_encoder}
\alias{save_token_text_encoder}
\title{Saves a token text encoder to a file}
\usage{
save_token_text_encoder(token_text_encoder, path)
}
\arguments{
\item{token_text_encoder}{a Token Text Encoder created with `token_text_encoder()`}

\item{path}{path to save the text encoder.}
}
\description{
Saves a token text encoder to a file
}
