% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_text.R
\name{subword_text_encoder}
\alias{subword_text_encoder}
\title{Subword Text Encoder}
\usage{
subword_text_encoder(vocab_list)
}
\arguments{
\item{vocab_list}{list of subwords for the vocabulary. Note that an underscore
at the end of a subword indicates the end of the word (i.e. a space will be
inserted afterwards when decoding). Underscores in the interior of subwords
are disallowed and should use the underscore escape sequence.}
}
\description{
Invertible TextEncoder using word pieces with a byte-level fallback.
}
\seealso{
\code{\link[=save_subword_text_encoder]{save_subword_text_encoder()}}, \code{\link[=load_subword_text_encoder]{load_subword_text_encoder()}}, \code{\link[=encode]{encode()}}
and \code{\link[=decode]{decode()}}
}
