% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{tfds}
\alias{tfds}
\title{TensorFlow Datasets module.}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 6.
}
\usage{
tfds
}
\description{
TensorFlow Datasets module.
}
\keyword{datasets}
