% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numpy.R
\name{tfds_as_numpy}
\alias{tfds_as_numpy}
\title{Converts a TensorFlow Dataset to an iterable of NumPy arrays.}
\usage{
tfds_as_numpy(dataset, graph = NULL)
}
\arguments{
\item{dataset}{a possibly nested structure of tf.data.Datasets and/or
tf.Tensors.}

\item{graph}{optional, explicitly set the graph to use.}
}
\description{
Converts a TensorFlow Dataset to an iterable of NumPy arrays.
}
