% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{tfds_load}
\alias{tfds_load}
\title{Loads the named dataset into a TensorFlow Dataset}
\usage{
tfds_load(name, split = NULL, data_dir = NULL, batch_size = NULL,
  download = TRUE, shuffle_files = FALSE, as_supervised = FALSE,
  decoders = NULL, builder_kwargs = NULL,
  download_and_prepare_kwargs = NULL, as_dataset_kwargs = NULL,
  try_gcs = FALSE, ...)
}
\arguments{
\item{name}{\code{str}, the registered name of the \code{DatasetBuilder} (the snake case
version of the class name). This can be either \code{"dataset_name"} or
\code{"dataset_name/config_name"} for datasets with \code{BuilderConfig}s.
As a convenience, this string may contain comma-separated keyword
arguments for the builder. For example \code{"foo_bar/a=True,b=3"} would use
the \code{FooBar} dataset passing the keyword arguments \code{a=True} and \code{b=3}
(for builders with configs, it would be \code{"foo_bar/zoo/a=True,b=3"} to
use the \code{"zoo"} config and pass to the builder keyword arguments \code{a=True}
and \code{b=3}).}

\item{split}{\code{tfds.Split} or \code{str}, which split of the data to load. If None,
will return a \code{dict} with all splits (typically \code{tfds.Split.TRAIN} and
\code{tfds.Split.TEST}).}

\item{data_dir}{\code{str} (optional), directory to read/write data.
Defaults to "~/tensorflow_datasets".}

\item{batch_size}{\code{int}, if set, add a batch dimension to examples. Note that
variable length features will be 0-padded. If
\code{batch_size=-1}, will return the full dataset as \code{tf.Tensor}s.}

\item{download}{\code{bool} (optional), whether to call
\code{tfds.core.DatasetBuilder.download_and_prepare}
before calling \code{tf.DatasetBuilder.as_dataset}. If \code{False}, data is
expected to be in \code{data_dir}. If \code{True} and the data is already in
\code{data_dir}, \code{download_and_prepare} is a no-op.}

\item{shuffle_files}{\code{bool}, whether to shuffle the input files.
Defaults to \code{False}.}

\item{as_supervised}{\code{bool}, if \code{True}, the returned \code{tf.data.Dataset}
will have a 2-tuple structure \code{(input, label)} according to
\code{builder.info.supervised_keys}. If \code{False}, the default,
the returned \code{tf.data.Dataset} will have a dictionary with all the
features.}

\item{decoders}{Nested dict of \code{Decoder} objects which allow to customize the
decoding. The structure should match the feature structure, but only
customized feature keys need to be present. See
\href{https://github.com/tensorflow/datasets/tree/master/docs/decode.md}{the guide}
for more info.}

\item{builder_kwargs}{\code{dict} (optional), keyword arguments to be passed to the
\code{tfds.core.DatasetBuilder} constructor. \code{data_dir} will be passed
through by default.}

\item{download_and_prepare_kwargs}{\code{dict} (optional) keyword arguments passed to
\code{tfds.core.DatasetBuilder.download_and_prepare} if \code{download=True}. Allow
to control where to download and extract the cached data. If not set,
cache_dir and manual_dir will automatically be deduced from data_dir.}

\item{as_dataset_kwargs}{\code{dict} (optional), keyword arguments passed to
\code{tfds.core.DatasetBuilder.as_dataset}.}

\item{try_gcs}{\code{bool}, if True, tfds.load will see if the dataset exists on
the public GCS bucket before building it locally.}

\item{...}{Additional parameters, currently not used.}
}
\description{
Loads the named dataset into a TensorFlow Dataset
}
