% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_text.R
\name{load_byte_text_encoder}
\alias{load_byte_text_encoder}
\title{Loads a Token Text Encoder}
\usage{
load_byte_text_encoder(path)
}
\arguments{
\item{path}{path of a saved token text encoder.}
}
\description{
Loads a Token Text Encoder
}
\seealso{
\code{\link[=byte_text_encoder]{byte_text_encoder()}}
}
